#58 [REFACTOR] DTO 패키지 구조 변경(request/response)
